feat: tap to change screen
