+++ androidApp/src/main/java/com/bluhabit/blu/android/presentation/authentication/onboard/OnboardScreen.kt
@@ -9,7 +9,6 @@ package com.bluhabit.blu.android.presentation.authentication.onboard
 
 import android.util.Log
 import androidx.activity.compose.rememberLauncherForActivityResult
-import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
@@ -19,16 +18,10 @@ import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.safeDrawingPadding
-import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.pager.HorizontalPager
 import androidx.compose.foundation.pager.rememberPagerState
-import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.material.Divider
-import androidx.compose.material.Icon
-import androidx.compose.material.LinearProgressIndicator
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
@@ -38,12 +31,9 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Brush
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.StrokeCap
 import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
@@ -54,6 +44,7 @@ import androidx.navigation.compose.rememberNavController
 import com.bluehabit.core.ui.R
 import com.bluhabit.blu.android.presentation.authentication.onboard.screen.FirstOnboardScreen
 import com.bluhabit.blu.android.presentation.authentication.onboard.screen.FourthOnboardScreen
+import com.bluhabit.blu.android.presentation.authentication.onboard.screen.ScreenFrameOnBoard
 import com.bluhabit.blu.android.presentation.authentication.onboard.screen.SecondOnboardScreen
 import com.bluhabit.blu.android.presentation.authentication.onboard.screen.ThirdOnboardScreen
 import com.bluhabit.blu.data.common.Response
@@ -64,9 +55,10 @@ import com.bluhabit.core.ui.theme.CustomTypography
 import com.bluhabit.core.ui.theme.UwangColors
 import com.bluhabit.core.ui.theme.UwangDimens
 import com.bluhabit.core.ui.theme.UwangTheme
-import com.bluhabit.core.ui.theme.UwangTypography
+import kotlinx.coroutines.coroutineScope
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flowOf
+import kotlinx.coroutines.launch
 
 val onboard = listOf(
     Triple(
@@ -117,6 +109,19 @@ fun OnboardScreen(
 
         else -> Modifier.background(MaterialTheme.colors.surface)
     }
+
+    suspend fun nextScreen() {
+        coroutineScope {
+            pagerState.scrollToPage(pagerState.currentPage + 1)
+        }
+    }
+
+    suspend fun prevScreen() {
+        coroutineScope {
+            pagerState.scrollToPage(pagerState.currentPage - 1)
+        }
+    }
+
     val googleAuthLauncher = rememberLauncherForActivityResult(contract = GoogleAuthContract(), onResult = {
         when (it) {
             is Response.Error -> {
@@ -158,10 +163,23 @@ fun OnboardScreen(
             },
             currentPage = pagerState.currentPage,
             skipOnboard = { },
+            nextScreen = {
+                scope.launch {
+                    pagerState.scrollToPage(
+                        pagerState.currentPage + 1
+                    )
+                }
+            },
+            prevScreen = {
+                scope.launch {
+                    pagerState.scrollToPage(
+                        pagerState.currentPage - 1
+                    )
+                }
+            },
             content = {
                 HorizontalPager(
                     state = pagerState,
-                    userScrollEnabled = true
                 ) { page ->
                     when (page) {
                         0 -> FirstOnboardScreen()
@@ -175,100 +193,6 @@ fun OnboardScreen(
     }
 }
 
-@Composable
-fun ScreenFrameOnBoard(
-    modifier: Modifier = Modifier,
-    headerTextColor: Color = UwangColors.Text.Secondary,
-    indicatorColor: Color = UwangColors.State.Primary.Main,
-    content: @Composable () -> Unit = {},
-    skipOnboard: () -> Unit = {},
-    currentPage: Int = 0,
-) {
-    val ctx = LocalContext.current
-    val dimens = UwangDimens.from(ctx)
-    Column(modifier = modifier.safeDrawingPadding()) {
-        Row(
-            modifier = Modifier.padding(dimens.dp_16),
-            horizontalArrangement = Arrangement.spacedBy(dimens.from(4.dp))
-        ) {
-            LinearProgressIndicator(
-                progress = 1f,
-                modifier = Modifier
-                    .weight(1f)
-                    .height(dimens.from(4.dp)),
-                color = indicatorColor,
-                strokeCap = StrokeCap.Round
-            )
-            LinearProgressIndicator(
-                progress = if (currentPage >= 1)  1f  else 0f,
-                modifier = Modifier
-                    .weight(1f)
-                    .height(dimens.from(4.dp)),
-                color = indicatorColor,
-                strokeCap = StrokeCap.Round,
-            )
-            LinearProgressIndicator(
-                progress = if (currentPage >= 2)  1f  else 0f,
-                modifier = Modifier
-                    .weight(1f)
-                    .height(dimens.from(4.dp)),
-                color = indicatorColor,
-                strokeCap = StrokeCap.Round
-            )
-            LinearProgressIndicator(
-                progress = if (currentPage >= 3)  1f  else 0f,
-                modifier = Modifier
-                    .weight(1f)
-                    .height(dimens.from(4.dp)),
-                color = indicatorColor,
-                strokeCap = StrokeCap.Round
-            )
-        }
-
-        Row(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(
-                    horizontal = dimens.dp_16
-                ),
-            horizontalArrangement = Arrangement.spacedBy(dimens.from(8.dp)),
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            Image(
-                painter = painterResource(id = R.drawable.app_logo),
-                contentDescription = "blu_logo",
-                modifier = Modifier
-                    .width(dimens.dp_24)
-                    .height(dimens.dp_24)
-            )
-            Text(
-                text = stringResource(id = R.string.label_header_logo),
-                style = UwangTypography.BodyMedium.Regular,
-                color = headerTextColor,
-            )
-            Spacer(modifier = Modifier.weight(1f))
-            Box(
-                modifier = Modifier
-                    .clip(CircleShape)
-                    .background(Color.White.copy(alpha = 0.2f))
-                    .padding(dimens.from(4.dp))
-            ) {
-                Icon(
-                    modifier = Modifier
-                        .clickable {
-                            skipOnboard()
-                        },
-                    painter = painterResource(id = R.drawable.ic_close),
-                    contentDescription = "ic_close",
-                    tint = headerTextColor
-                )
-
-            }
-
-        }
-        content()
-    }
-}
 
 @Composable
 fun StepFour(
@@ -277,7 +201,7 @@ fun StepFour(
     onNavigateToSignUp: () -> Unit,
     onNavigateToTermCondition: () -> Unit,
 
-) {
+    ) {
     ButtonGoogle(
         modifier = Modifier.fillMaxWidth(),
         text = stringResource(id = R.string.onboard_screen_sign_in_google_button_text),
